+++ source/split_multi_episode/changelog.md
@@ -1,4 +1,8 @@
 
+**<span style="color:#56adda">0.0.2</span>**
+- cache the results from silence detection so the 2 step silence detection doesn't have to rerun the detection a 2nd time.
+- remove TaskDataStore use and use the data object to modify the file movement behavior
+
 **<span style="color:#56adda">0.0.1</span>**
 
 - Initial release